# Switch from pkg_resources to importlib.metadata