--- a/setup.cfg
+++ b/setup.cfg
@@ -48,6 +48,10 @@ include =
 console_scripts =
     pygmentize = pygments.cmdline:main
 
+[options.extras_require]
+plugins =
+    importlib-metadata;python_version<'3.8'
+
 [aliases]
 release = egg_info -Db ''
 upload = upload --sign --identity=36580288