@Jean-Abou-Samra 2022-05-29
Bah. This failed the CI initially because `entry_points()` didn't accept the `group` keyword argument. Turns out, it was added in Python 3.10 and in recent versions of the Python <3.8 backport. Of course, I had tested those two versions, and not 3.8 or 3.9…